Compute the total price if the buyer buys any number of items with any cost. All items are same. There is a 10% discount if the price is more than 1000.
First, enter the input of the number of items. Then, enter the input of the cost of each item (since all the items are the same). 
Now, we know that the cost will be: cost of 1 item* total items. Hence, we have written that.
The discount will always be 10% of the cost. I wrote it in a different way but it means the same. Final Cost after discount is "cost-discount".
Then we write the final statements as I have written.
At last, we put a simple "Goodbye!" to look good. :)